\name{gubell}
\alias{gubell}

\docType{data}
\title{Guttman-Bell dataset}
\description{
  This dataset deals with attitudes of social groups.
}
\usage{gubell}
\format{
  The data frame consists of 7 objects and 5 variables with a total of 17 categories.
}

\references{
Guttman, L. (1968), A General Nonmetric Technique for Fitting the Smallest Coordinate Space for a Configuration
of Points, Psychometrika, 33, 469-506.
}


\examples{
   data(gubell)
   gubell
}
\keyword{datasets}