\name{hartigan}
\alias{hartigan}
\docType{data}
\title{Hartigan Data}
\description{
  Hartigan Data
}
\usage{data(hartigan)}
\format{
  A data frame with 24 observations on the following 6 variables.
  \describe{
    \item{thread}{a factor with levels \code{N} \code{Y}}
    \item{head}{a factor with levels \code{C} \code{F} \code{O} \code{R} \code{Y}}
    \item{indentation}{a factor with levels \code{L} \code{N} \code{T}}
    \item{bottom}{a factor with levels \code{F} \code{S}}
    \item{length}{a numeric vector}
    \item{brass}{a factor with levels \code{N} \code{Y}}
  }
}
\source{
  Clustering Algorithms, New York, Wiley, 1975, Table 12.7
}
\references{
  J.A. Hartigan
}
\examples{
data(hartigan)
}
\keyword{datasets}
