\name{house}
\alias{house}

\docType{data}
\title{House}
\description{
The votes selected cover a full spectrum of domestic, foreign, economic,
military, environmental and social issues. We tried to select votes which
display sharp liberal/conservative contrasts. In many instances we have chosen
procedural votes: amendments, motions to table, or votes on rules for debate. Often
these votes reveal true attitudes frequently obscured in the final votes.
}
\usage{house}
\format{
A data frame with the following variables:

Party: R = Republican, D = Democrat, I = Independent.

V1: HR 333. Bankruptcy Overhaul. Jackson-Lee (D-TX) amendment to allow
debtors to deduct additional medical and child-care expenses before
determining their eligibility for Chapter 7 bankruptcy status. The amendment
also expands the definition of family farmer, changes the standards
for calculating median income, and includes debtor privacy provisions. Rejected
160-258. March 1, 2001. A yes vote is a +.

V2: SJ Res 6. Ergonomics Rule Disapproval. Passage of the joint resolution
to reverse the ergonomics workplace safety rule submitted by the Clinton
Administration's Labor Department. Passed 223-206. March 7, 2001. A
no vote is a +.

V3: HR 3. Income Tax Reduction. Passage of the White House's bill to lower
federal income taxes by restructuring the five existing tax brackets into four
- 10 percent, 15 percent, 25 percent and 33 percent. The benefits of this tax
cut go disproportionately to the wealthy and to major corporations. The
large cost of the legislation would jeopardize domestic spending programs
aimed at middle- and low-income Americans. Passed 230-198. March 8,
2001. A no vote is a +.

V4: HR 6. Marriage Tax Reduction. Rangel (D-NY) substitute amendment
to reduce taxes by \$585.5 billion through 2011. This tax cut would be considerably
less regressive and more equitable than the Republican version.
The Rangel plan would create a new 12 percent bracket for the first \$20,000
of a couple's taxable income and \$10,000 for single taxpayers. It also
would increase the standard deduction for married couples filing jointly to
twice that of individuals filing singly. Additionally, the amendment would
simplify and expand the earned-income tax credit for low-income earners.
Rejected 196-231. March 29, 2001. A yes vote is a +.

V5: HR 8. Estate Tax Relief. Rangel (D-NY) substitute amendment to increase
the estate tax exemption from \$675,000 to \$2 million (\$4 million
for married couples) in 2002, rising to \$2.5 million by 2010. This legislation
serves as an alternative to the drastic Republican abolition of the
progressive estate tax. The Rangel tax cut would lower federal revenue
by \$39.2 billion over ten years. The amendment would retain current-law
step-up basis provisions, and replace the credit for estate taxes paid to a
state with a deduction. Rejected 201-227. April 4, 2001. A yes vote is a +.

V6: HR 503. Fetal Protection. Passage of the bill to make it a criminal offense
to injure or kill a fetus during the commission of a violent federal crime.
The measure would establish criminal penalties equal to those that would
apply if the injury or death occurred to a pregnant woman, regardless of the
perpetrator's knowledge of the pregnancy or intent to harm the fetus. The
bill states that its provisions should not be interpreted to apply to consensual
abortion or to a woman's actions with respect to her pregnancy. The
death penalty could not be imposed under this bill. Passed 252-172. April
26, 2001. A no vote is a +.

V7: HR 1. School Vouchers. Armey (R-TX) amendment to provide federal
funding for students to attend private schools, including religious schools,
if they are currently enrolled in schools that are dangerous or have been
low-performing for three years. Crime victims also would be provided
with funding to attend alternative private schools. Rejected 155-273. May
23, 2001. A no vote is a +.

V8: HR 1836. Tax Cut Reconciliation Bill. Adoption of the conference report
on the bill to reduce taxes by \$1.35 trillion through fiscal 2011 via income
tax rate cuts, relief of the marriage penalty, phaseout of the federal estate
tax, doubling of the child tax credit, and new incentives for retirement
savings. A new 10 percent tax rate would be created retroactive to January
1. The bill would: double the \$500-per-child tax credit by 2010 and make
it refundable; raise the estate tax exemption to \$1 million in 2002 and
repeal the tax in 2010; increase the standard deduction for married couples
to double that of singles over five years, beginning in 2005; and increase
annual contributions limits for Individual Retirement Accounts. The bill's
provisions would expire December 31, 2010. Adopted 240-154. May 26,
2001. A no vote is a +.

V9: HR 2356. Campaign Finance Reform. Adoption of the rule to allow the
House to consider a ban on soft money donations to national political
parties. This rule was crafted by campaign finance reform foes to disallow
amendments which fine-tune the bill and, thus, keep reform advocates
from gathering more votes in support of final passage. Beyond banning
soft money, the original reform legislation would allow up to \$10,000 in
soft-money donations to state and local parties for voter registration and
get-out-the vote activity. The reform bill would prevent issue ads from targeting
specific candidates within 60 days of a general election or 30 days
of a primary. Additionally, the legislation would maintain the current individual
contribution limit of \$1,000 per election for House candidates but
raise it to \$2,000 for Senate candidates, both of which would be indexed
for inflation. Rejected 203-228. July 12, 2001. A no vote is a +.

V10: HJ Res 36. Flag Desecration. Passage of the joint resolution proposing
a Constitutional amendment to prohibit physical desecration of the U. S.
flag. Passed 298-125. (A two-thirds majority vote of those present and
voting - 282 in this case - is required to pass a joint resolution proposing
an amendment to the Constitution.) July 17, 2001. A no vote is a +.

V11: HR 7. Faith-Based Initiative. Conyers (D-MI) motion to recommit the
bill to the Judiciary Committee with instructions to add language stating
that federally-funded religious service providers cannot discriminate based
on religion and that no provision supercedes state or local civil rights laws.
Motion rejected 195-234. July 19, 2001. A yes vote is a +.

V12: HJ Res 50. China Normalized Trade Relations. Passage of a joint resolution
to deny the President's request to provide normal trade relations
(formerly known as most-favored-nation trade status) for items produced
LOGIT AND PROBIT PCA 25
in China from July 2001 through July 2002. Rejected 169-259. July 19,
2001. A yes vote is a +.

V13: HR 4. ANWR Drilling Ban. Markey (D-MA) amendment to maintain the
current prohibition on oil drilling in the Arctic National Wildlife Refuge.
Rejected 206-223. August 1, 2001. A yes vote is a +.

V14: HR 2563. Patients' Rights/HMO Liability. Norwood (R-GA) amendment
to limit liability and damage awards when a patient is harmed by
denial of health care. This amendment was offered after patients' rights
opponents in the White House exerted pressure on Rep. Norwood to abandon
a stronger bill. The legislation would allow a patient to sue a health
maintenance organization (HMO) in state court but with federal, not state,
law governing. An employer could remove cases to federal court. The
bill would limit non-economic damages to \$1.5 million. Punitive damages
would be limited to the same amount and only allowed when a decisionmaker
fails to abide by a grant of benefits by an independent medical reviewer.
Adopted 218-213. August 2, 1001. A no vote is a +.

V15: HR 2563. Patients' Bill of Rights. Passage of the bill to provide federal
health care protections, such as access to specialty and emergency room
care, and require that health maintenance organizations (HMOs) have an
appeals process for patients who are denied care. This weakened legislation
was offered to head off consideration of a stronger version. A patient
denied care could sue an HMO in state and federal court but first must exhaust
internal and external appeals processes. Passed 226-203. August 2,
2001. A no vote is a +.

V16: HR 2944. Domestic Partner Benefits. Weldon (R-FL) amendment to
the FY 2002 District of Columbia Appropriations Bill that would prohibit
the use of local, as well as federal, funds to extend city employees' health
benefits to unmarried domestic partners. Rejected 194-226. September 25,
2001. A no vote is a +.

V17: HR 2586. U.S. Military Personnel Overseas/Abortions. Sanchez (DCA)
amendment to the FY 2002 Defense Authorization Bill which allows
female military personnel stationed at U.S. bases overseas to undergo an
abortion at medical facilities there provided they pay for it themselves and
a doctor consents to perform the operation. Rejected 199-217. September
25, 2001. A yes vote is a +.

V18: HR 2975. Anti-Terrorism Authority. Adoption of the rule to provide
for House consideration of the bill that would expand law enforcement's
power to investigate suspected terrorists and beef up domestic surveillance.
The legislation threatens the civil liberties, civil rights, and due process
protections guaranteed individuals in the United States. Adopted 214-208.
October 12, 2001. A no vote is a +.

V19: HR 3090. Economic Stimulus. Passage of the Republican version of the
post- September 11 economic stimulus package. The bill would grant businesses
and individuals \$99.5 billion in federal tax cuts in fiscal 2002, and
a total of \$159.4 billion in reductions over 10 years. Additionally, the bill
would allow more individuals to receive tax rebates for 2000, accelerate
a reduction of the 27 percent tax bracket to 25 percent, lower the capital
gains tax rate from 20 percent to 18 percent and eliminate the corporate
alternative minimum tax. Also, the legislation would provide \$3 billion to
states for health insurance for the unemployed. Passed 216-214. October
24, 2001. A no vote is a +.

V20: HR 3000. Trade Promotion Authority/Fast Track. Passage of the bill to
allow expedited negotiation and implementation of trade agreements between
the executive branch and foreign countries. The bill includes provisions
requiring increased consultations with Congress on any proposed
changes of tariffs for imports of sensitive agriculture products and on trade
disparities for textile products. Passed 215-214. December 6, 2001. A no
is a +.

}

\references{
Ada (2002). 2001 Voting Record: Shattered Promise of Liberal Progress. ADA Today, 57
(1), 1-17.
}


\examples{
   data(house)
   house
}
\keyword{datasets}
