\name{mammals}
\alias{mammals}
\docType{data}
\title{mammals}
\description{
  List of mammals categorized by dental properties.
}
\usage{data(mammals)}
\format{
  A data frame with 66 observations on the following 8 variables.
  \describe{
    \item{topincisors}{a numeric vector}
    \item{bottomincisors}{a numeric vector}
    \item{topcanines}{a numeric vector}
    \item{bottomcanines}{a numeric vector}
    \item{toppremolars}{a numeric vector}
    \item{bottompremolars}{a numeric vector}
    \item{topmolars}{a numeric vector}
    \item{bottommolars}{a numeric vector}
  }
}
\source{
  Clustering Algorithms, New York, Wiley, 1975, Table 9.1
}
\references{
  J.A. Hartigan
}
\examples{
data(mammals)
}
\keyword{datasets}
