\name{neumann}
\alias{neumann}

\docType{data}
\title{Neumann dataset}
\description{Willard Gibbs discovered a theoretical formula connecting the density, the
pressure, and the absolute temperature of a mixture of gases with convertible 
components. He applied this formula and the estimated constants to 65 experiments
of Neumann, and he discusses the systematic and accidental divergences (residuals).
}
\usage{neumann}
\format{A data frame with temperature, pressure, and density on 65 experiments.
}

\references{
Gifi, A. (1990). Nonlinear multivariate analysis. New York: Wiley.
Wilson, E.B. (1926). Empiricism and rationalism. Science, 64, 47-57
}


\examples{
   data(neumann)
   neumann
}
\keyword{datasets}
