\name{roskam}
\alias{roskam}

\docType{data}
\title{Roskam dataset}
\description{The Department of Psychology at the University of Nijmegen had 9 different
areas of research and teaching. Each of the 39 psychologists working in the
department ranked all 9 areas in order of relevance for their work. 
}
\usage{roskam}
\format{A data frame with the following 9 areas:

SOC: Social Psychology

EDU: Educational and Developmental Psychology

CLI: Clinical Psychology

MAT: Mathematical Psychology and Psychological Statistics

EXP: Experimental Psychology

CUL: Cultural Psychology and Psychology of Religion

IND: Industrial Psychology

TST: Test Construction and Validation

PHY: Physiological and Animal Psychology
}

\references{
Roskam, E (1968). Metric Analysis of Ordinal Data in Psychology. PhD
thesis, University of Leiden.
}


\examples{
   data(roskam)
   roskam
}
\keyword{datasets}
