\name{senate}
\alias{senate}

\docType{data}
\title{Senate votes}
\description{
2001 senate votes on 20 issues selected by Americans
for Democratic Action. The votes selected cover a full spectrum of domestic, foreign, economic,
military, environmental and social issues. We tried to select votes which
display sharp liberal/conservative contrasts. In many instances we have chosen
procedural votes: amendments, motions to table, or votes on rules for debate. Often
these votes reveal true attitudes frequently obscured in the final votes.
}
\usage{senate}
\format{
A data frame with the following variables:

Party: R = Republican, D = Democrat, I = Independent.

V1: Ashcroft Attorney General Confirmation. Confirmation of President
Bush's nomination of John Ashcroft of Missouri to serve as U.S. Attorney
General. Confirmed 58-42. Feb. 1, 2001. A no vote is a +.

V2: SJ Res 6. Ergonomics Rule Disapproval. Passage of a joint resolution
to reverse the ergonomics workplace safety rule submitted by the Clinton
Administration's Labor Department. Passed 56-44. March 6, 2001. A no
vote is a +.

V3: S 420. Social Security Lockbox. Domenici (R-NM) motion to waive
the Budget Act in order to ensure that the Social Security surplus is used
only to pay down the public debt until Social Security reform legislation is
enacted. The bill would also ensure that the surplus in the Medicare Hospital
Insurance Trust Fund is used only to pay down the public debt until
Medicare reform legislation is enacted. Motion rejected 52-48 (a threefifths
majority vote - 60 - is required to waive the Budget Act.) March 13,
2001. A no vote is a +.

V4: S 27. Campaign Finance Reform. McCain (R-AZ) motion to kill the
Hatch (R-UT) amendment requiring unions and corporations to obtain permission
from individual dues-paying workers or shareholders before spending
money on political activities. The Hatch amendment was intended as
a poison pill that, if passed and attached to the campaign finance reform
bill, would destroy any chances the full reform bill had of passage. The
Hatch amendment would also require corporations and unions to disclose
information regarding the funds spent on political activities. Motion agreed
to 69-31. March 21, 2001. A yes vote is a +.

V5: S 27. Soft Money Cap. McCain (R-AZ) motion to kill a Hagel (R-NE)
amendment to limit at \$60,000 per year soft money contributions by individuals,
political action committees, corporations and unions to national
and state political party committees. The Hagel amendment would render
the underlying reform bill's ban on soft money ineffective. Motion agreed
to 60-40. March 27. 2001. A yes vote is a +.

V6: H Con Res 83. Prescription Drug Benefit/Tax Cuts. Grassley (R-IA)
amendment to reserve \$300 billion over 10 years to create a Medicare prescription
drug benefit and overhaul the program. This amendment was a
response to Democratic legislation which would have allocated \$311 billion
for the benefit and not allowed the benefit's funding to come from the
Medicare Hospital Trust Fund Surplus. Adopted 51-50, with Vice President
Cheney casting a yea vote. April 3, 2001. A no vote is a +.

V7: H Con Res 83. Fiscal 2002 Budget Reconciliation. Domenici (R-NM)
amendment to instruct the Senate Finance Committee to report two reconciliation
bills to the Senate that would reduce revenue levels by not more
than the President's proposed \$1.6 trillion tax cut, and include a \$60 billion
economic stimulus package for fiscal 2001. Adopted 51-49. April 5, 2001.
A no vote is a +.

V8: H Con Res 83. Funding for Environmental Programs. Corzine (DNJ)
amendment to increase funding for a wide variety of environmental
programs by \$50 billion and set aside \$50 billion for debt reduction. The
increases would be offset by reductions in the proposed tax cut. Rejected
46-54. April 5, 2001. A yes vote is a +.

V9: H Con Res 83. Marriage Penalty Tax. Hutchison (R-TX) amendment
to increase the proposed tax cut by \$69 billion for fiscal 2002-2011 in an
effort to eliminate the co-called marriage penalty. Adopted 51-50, with
Vice President Cheney casting a yea vote. April 5, 2001. A no vote is a
+.

V10: H Con Res 83. Disabilities Education Act Funding. Breaux (D-LA)
amendment to redirect \$70 billion from the proposed tax cut to funding
for the Individuals with Disabilities Education Act (IDEA) over 10 years.
Adopted 54-46. April 5, 2001. A yes vote is a +.

V11: S I. School Renovation and Construction. Harkin (D-IA) amendment to
authorize \$1.6 billion for fiscal 2002 and such sums as necessary for each
fiscal year between 2003 and 2006 for the construction and renovation of
public elementary and secondary school buildings. Rejected 49-50. May
16, 2001. A yes vote is a +.

V12: HR 1836. Estate Tax. Dorgan (D-ND) amendment to strike the estatetax
repeal provision and repeal the estate tax in 2003 for only all qualified
family-owned farms and businesses. It also would reduce the top estate-tax
rate bracket to 45 percent. Rejected 43-56. May 21, 2001. A yes vote is a
+.

V13: HR 1836. Head Start. Kennedy (D-MA) amendment to condition the
reductions in the marginal income-tax rate on full funding for Head Start
programs. Motion rejected 45-54. May 22, 2001. A yes vote is a +.


V14: HR 1836. Tax Cut Reconciliation Bill. Adoption of the conference report
on the bill to reduce taxes by \$1.35 trillion through fiscal 2011 through income
tax rate cuts, relief of the marriage penalty, phase-out of the federal
estate tax, doubling of the child tax credit, and new incentives for retirement
savings. A new 10 percent tax rate would be created retroactive to
January 1. The bill would double the \$5000-per-child tax credit by 2010
and make it refundable, raise the estate tax exemption to \$1 million in 2002
and repeal the tax in 2010, increase the standard deduction for married
couples to double that of singles over five years, beginning in 2005, and
increase annual contributions limits for Individual Retirement Accounts.
The bill's provisions would expire December 31, 2010. Adopted 58-33.
May 26, 2001. A no vote is a +.

V15: S I. School Vouchers. Gregg (R-NH) amendment to create a demonstration
program in 10 school districts to provide public school children with
federal funds (vouchers) to transfer to another public school or a private
school, including religious schools. The amendment would authorize \$50
million for fiscal 2002 and subsequent necessary sums for the next six fiscal
years. Rejected 41-58. June 12, 2001. A no vote is a +.

V16: S I. Boy Scouts/Anti-Discrimination. Helms (R-NC) amendment to withhold
federal education funds from public elementary and secondary schools
that bar the Boy Scouts of America from using school facilities. The targeted
schools bar the Boy Scouts because the organization discriminates
against gay men. Adopted 51-49. June 14, 2001. A no vote is a +.

V17: S 1052. Patients' Bill of Rights. Passage of the bill to provide federal
patient protections and allow patients to appeal a health maintenance organization's
(HMO) decision on coverage and treatment. It also would allow
patients to sue health insurers in state courts over quality-of-care claims
and, at the federal level, over administrative or non-medical coverage disputes.
Passed 59-36. June 29, 2001. A yes vote is a +.

V18: HR 2299. NAFTA/Mexican Trucks. Shelby (R-AL) motion to uphold a
border truck inspection program which allows Mexican trucks to receive
three-month permits if they pass safety inspections. The motion also upholds
a grant of \$60 billion to the Transportation Department and various
agencies. Motion agreed to 65-30. July 27, 2001. A yes vote is a +.

V19: S 1438. Military Base Closures. Warner (R-VA) motion to authorize an
additional round of U.S. military base realignment and closures in 2003.
Motion agreed to 53-47. September 25, 2001. A yes vote is a +.

V20: HR 2944. Fiscal 2002 District of Columbia Appropriations. Passage of
the bill to provide \$408 million for the District of Columbia in fiscal 2002,
including funds for the city's courts and corrections system and \$16.1 million
for an emergency response plan following the September 11 attacks.
The bill also would approve a \$7.2 billion budget for the District. Passed
75-24. November 7, 2001. A yes is a +.
}

\references{
Ada (2002). 2001 Voting Record: Shattered Promise of Liberal Progress. ADA Today, 57
(1), 1-17.
}


\examples{
   data(senate)
   senate
}
\keyword{datasets}
