\name{senate07}
\alias{senate07}

\docType{data}
\title{Senate votes 2007}
\description{
2007 senate votes on 20 issues selected by Americans
for Democratic Action. The votes selected cover a full spectrum of domestic, foreign, economic,
military, environmental and social issues. We tried to select votes which
display sharp liberal/conservative contrasts. In many instances we have chosen
procedural votes: amendments, motions to table, or votes on rules for debate. Often
these votes reveal true attitudes frequently obscured in the final votes.
}
\usage{senate}
\format{
A data frame (98 senators) with the following variables:

Party: R = Republican, D = Democrat, I = Independent.

I1. S 1. Ethics and Lobbying Overhaul: Grass-Roots Lobbying Bennett (R-UT)
amendment to Reid (D-NV) ethics reform legislation. Reid's legislation would overhaul
Congressional lobbying and ethics rules for Senators and their staffs, restrict use of
more broadly defi ned earmarks, prohibit Senators and staff from accepting gifts
and free meals, extend the period before former Senators can become lobbyists,
prohibit lobbyist-funded travel, and require trips funded by private entities to be vetted
by the Select Ethics Committee. The Bennett amendment, permitting the travel
of Members of Congress hosted by 501 (c) (3), or 'grass-roots lobbying' organizations,
was adopted 51-46. Jan. 17, 2007. A no vote is a +.

I2. HR 2. Minimum Wage Increase: Motion to invoke cloture (thus limiting debate) on
a Baucus (D-MT) substitute amendment to raise the minimum wage to \$7.25 per hour
over two years and provide \$8.3 billion in small-business tax incentives. The tax
provisions would be offset with revenue increases, including a \$1 million cap on
the amount of executive compensation that can be tax-deferred in any year and extend
certain restrictions on certain sale-in-leaseout deals. (Three-fifths of the total Senate,
60 votes, is required to invoke cloture.) Motion agreed to 87-10. Jan 30, 2007. A
yes vote is a +.

I3. S 574. Iraq War 'Surge': Motion to invoke cloture and proceed to a bill that
would express the sense of Congress in support of U.S. military serving in Iraq,
while disapproving of President Bush's 'surge' decision to deploy more than
20000 additional U.S. combat troops to that country and requiring the President to
submit a report to Congress every 30 days on military progress in Iraq. Motion rejected
56-34. Feb. 17, 2007. A yes vote is a +.

I4. S J Res 9. Iraq Mission: Passage of a joint resolution to limit the mission of U.S.
forces in Iraq, within 120 days of enactment, to counter-terrorism efforts, training Iraqi
forces and protecting U.S. assets. It would  set a binding goal of withdrawing most
combat troops by March 31, 2008. (By unanimous consent, the Senate agreed to
require 60 votes to pass the joint resolution.) Rejected 48-50. March 15, 2007. A yes
vote is a +.

I5. S 3. Medicare Prescription Drug Price Negotiation: Motion to invoke cloture
and proceed to a bill allowing the Health and Human Services Department to negotiate
the prices drug companies charge under the Medicare Part D prescription drug program.
Motion rejected 55-42. April 18, 2007. A yes vote is a +.

I6. HR 1495. Water Resources Development
Act Reauthorization - Climate Change: Kerry (D-MA) amendment to Boxer (D-CA)
legislation authorizing about \$14 billion for Army Corps of Engineers water resources
development projects, including initiatives for storm protection in Louisiana, fl ood
control, and ecosystem restoration. The Kerry amendment would require the Army
Corps of Engineers to include the impact on global climate change in any relevant study
of water resource projects.(By unanimous consent, the Senate agreed to require 60
votes to adopt the amendment.) Rejected 51-42. May 15, 2007. A yes vote is a +.

I7. S 1348. Immigration Overhaul: Motion to invoke cloture on Kennedy (D-MA)
legislation to overhaul U.S. immigration policies and institute new border security
measures, including an electronic verification system. It would allocate some
green cards on a 'merit-based' system based on certain criteria, including a highdemand
occupation, profi ciency in English, and higher education degrees. Half of the
green cards would be allocated based on family relations. It also would provide for
a temporary guest worker program that would allow workers to remain in the United
States for up to six years, provided that they return to their home country for a year after
every two years they remain in the United States. A 'yea' was a vote in support of
Kennedy's and the President's position. Motion rejected 45-50. June 7, 2007. A yes
vote is a +.

I8. HR 800. EFCA: Motion to invoke cloture on the Reid (D-NV) motion to proceed to
a bill that would allow union organizers to bypass an NLRB-sponsored election
if a majority of eligible employees sign a petition in support of a union. Motion
rejected 51-48. June 26, 2007. A yes vote is a +.

I9. HR 1585. Fiscal 2008 Defense Authorization: Motion to invoke cloture
on a Levin (D-MI) amendment mandating a drawdown and redeployment of U.S.
forces in Iraq by the end of April 2008, except to protect U.S. personnel, train
Iraqi security forces, and conduct counter terrorism operations against al Qaeda and
other international terrorist organizations. Redeployment would begin within 120
days of enactment. The bill would authorize \$648.3 billion for defense programs in
fiscal 2008, including \$127.5 billion for the wars in Iraq and Afghanistan. It
also would authorize \$143.5 billion for operations and maintenance; \$109.9 billion
for procurement; \$122.9 billion for military personnel, and \$74.7 billion for research,
development, testing and evaluation. Motion rejected 52-47. July 18, 2007. A
yes vote is a +.

I10. S 2011. FISA: Passage of a bill to amend the Foreign Intelligence Surveillance
Act of 1978 (FISA) to expand the authority  of the Attorney General and the Director of
National Intelligence to conduct surveillance of communication between one suspected
foreign terrorist and another without a court warrant. Communications companies
would be compelled to comply with the government's conduct of such surveillance.
Within 10 days of initiating the electronic surveillance, the Administration would be
required to describe to the FISA court the procedures they used to determine whether
the intelligence acquisition being conducted without a warrant is directed at foreign
targets overseas. The FISA court would then rule on the procedures. The bill would
sunset after six months. (By unanimous  consent, the Senate agreed to require 60
votes to pass the bill.) Rejected 43-45. Aug. 3, 2007. A yes vote is a +.

I11. S 1257. District of Columbia Voting Rights Motion to invoke cloture on a Reid
(D-NV) motion to proceed to a bill to increase the membership of the House of
Representatives to 437, by granting a seat to the District of Columbia and an additional
seat to Utah. Motion rejected 57-42. Sept. 19, 2007. A yes vote is a +.

I12.HR 1585. Fiscal 2008 Defense Authorization: Motion to invoke cloture
on the Kennedy (D-MA) amendment to Levin (D-MI) legislation (see vote No. 9,
above). The Kennedy amendment would make violent crimes that cause bodily harm
based on the victim's race, color, religion or national origin punishable by a fi ne and up
to 10 years in prison, and punishable by a life sentence if the victim dies, is kidnapped
or subjected to aggravated sexual abuse. It also would create the same penalties
for crimes motivated by gender, sexual orientation, gender identity, or disability.
(Subsequently, the Kennedy amendment was adopted by voice vote.) Motion agreed
to 60-39. Sept. 2007. A yes vote is a +.

I13. HR 976. SCHIP: Motion to invoke cloture on the Reid (D-NV) motion to
agree to the House bill reauthorizing the State Children's Health Insurance Program
(SCHIP) at nearly \$60 billion over five years, expanding the program by \$35
billion. To offset the cost of the expansion, the bill would increase the tax on cigarettes
by 61 cents to \$1 per pack and raise taxes on other tobacco products. Motion agreed to
69-30. Sept. 27, 2007. A yes vote is a +.

I14. HR 3043. Fiscal 2008 Labor - HHS-Education Appropriations - Limiting Abortion Funding: 
Vitter (R-LA) amendment to Harkin (D-IA) legislation. The Vitter amendment would prohibit the
distribution of Title X family planning funds to abortion providers except when a woman
suffers from a physical disorder, injury, or illness that would risk the life of the
mother unless an abortion is provided. The amendment would not apply to hospitals.
Harkin's legislation would appropriate \$605.5 billion in fiscal 2008, including
\$149.9 billion in discretionary spending, for the departments of Labor, Health and
Human Services (HHS), and Education, and for related agencies. It would provide \$63
billion for the Education Department, \$14.9 billion for the Labor Department and \$479.1
billion for HHS. Rejected 41-52. Oct. 18, 2007. A no vote is a +.

I15. Southwick Nomination: Motion to invoke cloture on the nomination of Leslie
Southwick of Mississippi to be a judge for the 5th Circuit Court of Appeals. Opponents
argued that, in his positions related to African Americans, workers, gays, and
women, Mr. Southwick was unsuited to a lifetime appointment to the federal judiciary.
Motion agreed to 62-35. Oct. 24, 2007. A no vote is a +.

I16. HR 1495. Water Resources Development - Veto Override: Passage, over President 
Bush's Nov. 2, 2007, veto, of  a bill authorizing \$23.2 billion for more than
900 water resource development projects and studies by the Army Corps of Engineers
for flood control, navigation, beach erosion control, and environmental restoration. The
bill would require independent peer review for certain projects that exceed \$40 million.
It would authorize \$3.9 billion for a system of new locks and dams and environmental
restoration for the Upper Mississippi River and Illinois waterway system. It also would
authorize hurricane recovery activities along the Gulf Coast. (A two-thirds majority of
those present and voting of both chambers - 62 senators in this instance - is required
to override a veto.) Passed, thus enacted into law, 79-14. Nov. 8, 2007. A yes vote is
a +.

I17.HR 3688. U.S.-Peru Trade Agreement: Passage of a bill that would implement a
NAFTA/CAFTA model trade agreement between the United States and Peru. The
agreement would reduce most tariffs and duties currently affecting trade between
the two countries, increase protections for intellectual property and would require
Peru to take steps to strengthen its labor and environmental-enforcement standards.
Passed, thus cleared for the President, 77-18. Dec. 4, 2007. A no vote is a +.

I18. HR 3996. AMT Adjustment: Motion to invoke cloture on a Reid (D-NV) motion
to proceed to the bill that would provide a one-year adjustment to the alternative
minimum tax (AMT) on 2007 income, exempt an additional 21 million taxpayers
from paying the AMT, extend several expiring tax provisions for one year, and
expand eligibility in 2008 for the refundable child tax credit. To offset the cost, the bill
includes revenue raisers, such as increasing the tax on the carried interest of private
equity managers, venture capitalists and some real estate investors. Motion rejected
46-48. Dec. 6, 2007. A yes vote is a +.

I19. HR 6. Energy Policy: Motion to invoke cloture on the Reid (D-NV) motion to
concur in the House bill with an additional amendment to require new Corporate
Average Fuel Economy standards of 35 miles per gallon for cars and light trucks,
and require the production and use of 36 billion gallons of biofuels by 2022. It would
direct the Energy Department to set new energy efficiency standards. The additional
amendment would strike a provision requiring utilities to produce 15 percent of
their electricity from alternative sources by 2020. It also would increase to \$21.8 billion
a package of tax incentives that would be offset in part by eliminating or reducing
\$13 billion in subsidies for major oil and gas companies. Motion rejected 59-40. Dec.
13, 2007. A yes vote is a +.

I20. HR 2419. Farm Bill Reauthorization Passage of a bill to authorize \$286 billion
from fiscal 2008 through fiscal 2012 for federal farm, nutrition assistance,
rural development, and agricultural trade programs. Producers would not be eligible
if the average of their three-year adjusted gross income is more than \$1 million in
2009, and more than \$750,000 in 2010 and subsequent years, unless at least two-thirds
of that income is derived from agriculture. The bill also would limit the amount that
farms could receive during any crop year under specific programs to \$40,000 for
direct and fi xed payments, and \$60,000 for counter-cyclical and average crop revenue
payments. It would create a new optional subsidy called the Average Crop Revenue
program, which would tie payments to state based farm revenue figures. Country-of-origin
labels for all meat would be required by September 2008. The bill would extend
the Milk Income Loss Contract program through fiscal 2012. It would increase,
from \$134 to \$140 per month, the amount households could deduct from gross income
in determining their level of food stamp benefits. It also would authorize \$1.1
billion to promote farm-based renewable energy and effi ciency. (Before passage, the
Senate adopted a Harkin (D-IA) substitute amendment by voice vote.) Passed 79-14.
Dec. 14, 2007. A yes vote is a +.
}

\references{
Ada (2002). 2001 Voting Record: Shattered Promise of Liberal Progress. ADA Today, 57
(1), 1-17.
}


\examples{
   data(senate07)
   senate07
}
\keyword{datasets}
