\name{sleeping}
\alias{sleeping}

\docType{data}
\title{Sleeping Bags}
\description{
  This data set provides 4 variables measured on 21 sleeping bags. The variables are temperature,
  weight, price, and material.
}
\usage{sleeping}
\format{
  A data frame of dimenson 21 times 4.
}

\references{
Prediger, S. (1997). Symbolic objects in formal concept analysis. In G. Mineau, and 
A. Fall (eds.), Proceedings of the 2nd International Symposium on Knowledge, Retrieval,
Use, and Storage for Efficiency.
}

\examples{
   data(sleeping)
   sleeping
}
\keyword{datasets}