\name{small}
\alias{small}

\docType{data}
\title{Small dataset}
\description{
  A small artifical dataset.
}
\usage{small}
\format{
  A data frame with 3 variables each of them has 3 categories.
}

\references{
Gifi, A. (1990). Nonlinear Multivariate Analysis: Wiley.
}


\examples{
   data(small)
   small
}
\keyword{datasets}