\name{sleeping}
\alias{sleeping}

\docType{data}
\title{Sleeping Bags}
\description{
  Describtion of 21 sleeping bags.
}
\usage{sleeping}
\format{
  A data frame with three variables (price, filling and quality) with a total of 8 categories.
}

\references{
De Leeuw, J., and Michailidis, G. (2000). Graph Layout Techniques and Multidimensional Data Analysis, Game
Theory, Optimal Stopping, Probability and Statistics. Papers in honor of Thomas S. Ferguson, F.T. Bruss and L.
Le Cam (eds.), IMS Lecture Notes-Monograph Series, 219-248
}


\examples{
   data(sleeping)
   sleeping
}
\keyword{datasets}